+++ src/basic/modules/getStockTotal.js
@@ -1,0 +1,11 @@
+export function getStockTotal(prodList) {
+  let sum;
+  let i;
+  let currentProduct;
+  sum = 0;
+  for (i = 0; i < prodList.length; i++) {
+    currentProduct = prodList[i];
+    sum += currentProduct.q;
+  }
+  return sum;
+}
