+++ src/basic/modules/handleCartUpdate.js
@@ -1,0 +1,217 @@
+import { updateLoyaltyPointsDisplay } from './updateLoyaltyPointsDisplay.js';
+import { handleStockInfoUpdate } from './handleStockInfoUpdate.js';
+import { PRODUCT_IDS } from '../constants/business-rules.js';
+
+export function handleCartUpdate({ cartDisp, prodList }) {
+  let idx;
+  let originalTotal;
+  let savedAmount;
+  let points;
+  let previousCount;
+  let stockMsg;
+  let totalAmt = 0;
+  let itemCnt = 0;
+  originalTotal = totalAmt;
+  const cartItems = cartDisp.children;
+  let subTot = 0;
+  const itemDiscounts = [];
+  const lowStockItems = [];
+  for (idx = 0; idx < prodList.length; idx++) {
+    if (prodList[idx].q < 5 && prodList[idx].q > 0) {
+      lowStockItems.push(prodList[idx].name);
+    }
+  }
+  for (let i = 0; i < cartItems.length; i++) {
+    (() => {
+      let curItem;
+      for (let j = 0; j < prodList.length; j++) {
+        if (prodList[j].id === cartItems[i].id) {
+          curItem = prodList[j];
+          break;
+        }
+      }
+      const qtyElem = cartItems[i].querySelector('.quantity-number');
+      const q = Number.parseInt(qtyElem.textContent, 10);
+      const itemTot = curItem.val * q;
+      let disc = 0;
+      itemCnt += q;
+      subTot += itemTot;
+      const itemDiv = cartItems[i];
+      const priceElems = itemDiv.querySelectorAll('.text-lg, .text-xs');
+      priceElems.forEach(elem => {
+        if (elem.classList.contains('text-lg')) {
+          elem.style.fontWeight = q >= 10 ? 'bold' : 'normal';
+        }
+      });
+      if (q >= 10) {
+        if (curItem.id === PRODUCT_IDS.keyboard) {
+          disc = 10 / 100;
+        } else {
+          if (curItem.id === PRODUCT_IDS.mouse) {
+            disc = 15 / 100;
+          } else {
+            if (curItem.id === PRODUCT_IDS.monitorArm) {
+              disc = 20 / 100;
+            } else {
+              if (curItem.id === PRODUCT_IDS.laptopPouch) {
+                disc = 5 / 100;
+              } else {
+                if (curItem.id === PRODUCT_IDS.speaker) {
+                  disc = 25 / 100;
+                }
+              }
+            }
+          }
+        }
+        if (disc > 0) {
+          itemDiscounts.push({ name: curItem.name, discount: disc * 100 });
+        }
+      }
+      totalAmt += itemTot * (1 - disc);
+    })();
+  }
+  let discRate = 0;
+  originalTotal = subTot;
+  if (itemCnt >= 30) {
+    totalAmt = (subTot * 75) / 100;
+    discRate = 25 / 100;
+  } else {
+    discRate = (subTot - totalAmt) / subTot;
+  }
+  const today = new Date();
+  const isTuesday = today.getDay() === 2;
+  const tuesdaySpecial = document.getElementById('tuesday-special');
+  if (isTuesday) {
+    if (totalAmt > 0) {
+      totalAmt = (totalAmt * 90) / 100;
+      discRate = 1 - totalAmt / originalTotal;
+      tuesdaySpecial.classList.remove('hidden');
+    } else {
+      tuesdaySpecial.classList.add('hidden');
+    }
+  } else {
+    tuesdaySpecial.classList.add('hidden');
+  }
+  document.getElementById('item-count').textContent =
+    '🛍️ ' + itemCnt + ' items in cart';
+  const summaryDetails = document.getElementById('summary-details');
+  summaryDetails.innerHTML = '';
+  if (subTot > 0) {
+    for (let i = 0; i < cartItems.length; i++) {
+      let curItem;
+      for (let j = 0; j < prodList.length; j++) {
+        if (prodList[j].id === cartItems[i].id) {
+          curItem = prodList[j];
+          break;
+        }
+      }
+      const qtyElem = cartItems[i].querySelector('.quantity-number');
+      const q = Number.parseInt(qtyElem.textContent, 10);
+      const itemTotal = curItem.val * q;
+      summaryDetails.innerHTML += `
+        <div class="flex justify-between text-xs tracking-wide text-gray-400">
+          <span>${curItem.name} x ${q}</span>
+          <span>₩${itemTotal.toLocaleString()}</span>
+        </div>
+      `;
+    }
+    summaryDetails.innerHTML += `
+      <div class="border-t border-white/10 my-3"></div>
+      <div class="flex justify-between text-sm tracking-wide">
+        <span>Subtotal</span>
+        <span>₩${subTot.toLocaleString()}</span>
+      </div>
+    `;
+    if (itemCnt >= 30) {
+      summaryDetails.innerHTML += `
+        <div class="flex justify-between text-sm tracking-wide text-green-400">
+          <span class="text-xs">🎉 대량구매 할인 (30개 이상)</span>
+          <span class="text-xs">-25%</span>
+        </div>
+      `;
+    } else if (itemDiscounts.length > 0) {
+      itemDiscounts.forEach(item => {
+        summaryDetails.innerHTML += `
+          <div class="flex justify-between text-sm tracking-wide text-green-400">
+            <span class="text-xs">${item.name} (10개↑)</span>
+            <span class="text-xs">-${item.discount}%</span>
+          </div>
+        `;
+      });
+    }
+    if (isTuesday) {
+      if (totalAmt > 0) {
+        summaryDetails.innerHTML += `
+          <div class="flex justify-between text-sm tracking-wide text-purple-400">
+            <span class="text-xs">🌟 화요일 추가 할인</span>
+            <span class="text-xs">-10%</span>
+          </div>
+        `;
+      }
+    }
+    summaryDetails.innerHTML += `
+      <div class="flex justify-between text-sm tracking-wide text-gray-400">
+        <span>Shipping</span>
+        <span>Free</span>
+      </div>
+    `;
+  }
+  const totalDiv = document.querySelector('.text-2xl');
+  if (totalDiv) {
+    totalDiv.textContent = '₩' + Math.round(totalAmt).toLocaleString();
+  }
+  const loyaltyPointsDiv = document.getElementById('loyalty-points');
+  if (loyaltyPointsDiv) {
+    points = Math.floor(totalAmt / 1000);
+    if (points > 0) {
+      loyaltyPointsDiv.textContent = '적립 포인트: ' + points + 'p';
+      loyaltyPointsDiv.style.display = 'block';
+    } else {
+      loyaltyPointsDiv.textContent = '적립 포인트: 0p';
+      loyaltyPointsDiv.style.display = 'block';
+    }
+  }
+  const discountInfoDiv = document.getElementById('discount-info');
+  discountInfoDiv.innerHTML = '';
+  if (discRate > 0 && totalAmt > 0) {
+    savedAmount = originalTotal - totalAmt;
+    discountInfoDiv.innerHTML = `
+      <div class="bg-green-500/20 rounded-lg p-3">
+        <div class="flex justify-between items-center mb-1">
+          <span class="text-xs uppercase tracking-wide text-green-400">총 할인율</span>
+          <span class="text-sm font-medium text-green-400">${(discRate * 100).toFixed(1)}%</span>
+        </div>
+        <div class="text-2xs text-gray-300">₩${Math.round(
+          savedAmount
+        ).toLocaleString()} 할인되었습니다</div>
+      </div>
+    `;
+  }
+  const itemCountElement = document.getElementById('item-count');
+  if (itemCountElement) {
+    previousCount = Number.parseInt(
+      itemCountElement.textContent.match(/\d+/) || 0,
+      10
+    );
+    itemCountElement.textContent = '🛍️ ' + itemCnt + ' items in cart';
+    if (previousCount !== itemCnt) {
+      itemCountElement.setAttribute('data-changed', 'true');
+    }
+  }
+  stockMsg = '';
+  for (let stockIdx = 0; stockIdx < prodList.length; stockIdx++) {
+    const item = prodList[stockIdx];
+    if (item.q < 5) {
+      if (item.q > 0) {
+        stockMsg =
+          stockMsg + item.name + ': 재고 부족 (' + item.q + '개 남음)\n';
+      } else {
+        stockMsg = stockMsg + item.name + ': 품절\n';
+      }
+    }
+  }
+  const stockInfo = document.getElementById('stock-status');
+  stockInfo.textContent = stockMsg;
+  handleStockInfoUpdate({ prodList, stockInfo });
+  updateLoyaltyPointsDisplay({ cartDisp, prodList, totalAmt, itemCnt });
+}
