+++ src/basic/modules/index.js
@@ -1,0 +1,6 @@
+export * from './handleSelectOptionsUpdate.js';
+export * from './handleCartUpdate.js';
+export * from './updatePricesInCart.js';
+export * from './handleStockInfoUpdate.js';
+export * from './updateLoyaltyPointsDisplay.js';
+export * from './getStockTotal.js';
